Add destDirProblem and simplify ensureDestDir
This commit actually changes several things.

The old behavior was: when we started up the application, we'd check the "destination directory" that was specified in prefs.xml, and try to verify that it existed.  If it didn't exist, we'd try to create it.  We went through a number of different checks, and formulated different error messages.  If we couldn't create it, we would issue a specific warning via the logger, automatically disable "move", and have a dialog box telling the user that the directory was bad and that move has been disabled.

We ran that code from UIStarter, some time after we open up the Shell, but before we start the event loop.

With this commit, we still check the destination dir.  But first off, it's a simpler check, using FileUtilities.checkForCreatableDirectory.  This does not actually try to create the directory, it just determines if we COULD.  If we can't, we do NOT any longer disable move.  If "moveSelected" was true, it remains true, to indicate that the user WANTS us to move the files.  But we introduce a new transient variable, destDirProblem, and set it to true when checkForCreatableDirectory returns false.

Since we no longer disable "move", we no longer say that to the user.  Instead, we say that the move will not be possible.  This aspect still needs a little work going forward.

Now we're calling ensureDestDir from setMoveEnabled().  There was a longstanding comment there saying that maybe we should.  Now we *need* to, to make sure destDirProblem stays up to date.

This is important because we don't actually verify the directory when move is not selected.  If the startup config has move de-selected, we ignore the dest dir.  So if the user opens up the preferences dialog and selects "move", we need to verify the directory then.