Continue with earlier change: say "select", not "enable"
To indicate that the user had indicated they want files to be moved, we will say move is "selected".  To indicate they want files renamed, we say rename is "selected".  We'll try to use "enabled" for different things.

So we now have a UserPreferences method, "isMoveSelected".  That returns whether or not the user has selected to "move" the files.

But we also have "isMoveEnabled".  The difference is, "isMoveEnabled" suggests we are ready and able to move the files; i.e., the dest dir seems valid.  "isMoveSelected" just tells us the user preference.

Most clients should only be interested in "isMoveSelected".

For rename, there's no equivalent of "enabled".  "Selected" implies "enabled".

There's one other problem due to the use of XStream.  The variables have been written to the XML as "moveEnabled" and "renameEnabled".  If we simply change the variable names, the XML reading will fail while reading in the old names.  We can resolve this, though.  We establish an alias in UserPreferencesPersistence.  That applies to reading and writing.  So, even after loading the old file, opening the preferences dialog, and saving your changes, prefs.xml will STILL have "moveEnabled" and "renameEnabled". Hopefully it won't confuse anyone.