Move checkDestinationDirectory from UIStarter to ResultsTable
Now we call it when the UIStarter calls the ready() method of ResultsTable.  Also make the method synchronized.

It was probably a mistake that it didn't move to ResultsTable when we first broke out ResultsTable into its own file.  This was the only reason UIStarter had anything to do with UserPreferences, or in fact, anything to do with the model.  Now that we've moved this code, UIStarter doesn't depend on model at all (besides for Constants), and is very generic.  With a few small changes, it could be used as a startup file for any graphic application.

Also, I plan to use the results of checkDestinationDirectory specifically to change the results table, but that's still to be done.