Rework updateUserPreferences in ResultsTable
Change to a switch statement.  It's cleaner, and it allows you to omit the enum name in the case labels.  But beyond that, we actually want to take advantage of the most reviled "feature" of the switch statement: the fall through.  The fact that a "break" is required to prevent fall through is generally a source of bugs, but in our case, we want to fall through, and it's very convenient.

We are now calling setActionButtonText when the userPref is DEST_DIR; we used to not, but the dest dir is incorporated into the tooltip, so if it changes, we do need to update the button.