Do checkDestinationDirectory also for move/rename change
Call checkDestinationDirectory on move/rename change, in addition to dest dir change.  For example, if the user has set a bad directory as their dest dir, but has had "move" unselected, there never would have been a warning.  If they then select "move", we need to check the directory and generate the dialog box to tell the user what's going on.