Disable button when action makes no sense
We now will disable the action button if either of the following is true:
1) neither move nor rename is selected
2) move is selected, but the destination directory is not valid

Since the button might be disabled, we always enable the button at the top of setActionButtonText, and then possibly re-disable it if the situation still calls for it.

Note we are setting the tooltip even in the case where we disable the button.  Disabled controls don't provide tooltips.  But I hope to find a way arond that at some point (an overlay of some kind?).  So I'm leaving the tooltips in place.

(View the diffs with ?w=1 to ignore whitespace.)