Modify creation of TableItem and ComboBox
createTableItem does not need to take the table as an argument, it can get that using the instance variable.

And it obviously doesn't need to take the filename, either.  I think I had it that way because the caller used to need the filename for other stuff too, so I figured, as long as it was handy, just pass it as an argument.  But by now, the caller was only looking up the filename to send to createTableItem, so stop doing that.  Have createTableItem look up the filename itself.

Use new methods for instantiating TableItem and ComboBox objects.