Rearrange methods in ResultsTable.java
Try to have the manipulation methods at the top of the file, and the setup ones at the bottom.  Try to define methods before calling them.

This introduces an unfortunate amount of diffs, and makes things like "git blame" less useful, but it will make it a lot easier for me to merge in my changes from another branch I was working in.  And I do think it's better this way, too.