Make StringUtils.SANITISE an unmodifiable map
Use Collections.unmodifiableMap to make SANITISE not modifiable.  Actually, in the future, we might want it to be modifiable.  We might want to let users configure the substitution.  But for now, csince the functionality is not present, make it unmodifiable.

Add tests to ensure that the map is accessible, but not modifiable.

The testUnmodifiableMap() test does a try/catch with nothing but a comment in the catch block.  This has caused findbugs to complain, though bizarrely, not consistently.  Add an exclusion anyway, so it never comes up.