Declare updateUserPreferences switch statement safe
Tell findbugs that it's ok that ResultsTable.updateUserPreferences has switch cases which fall through.  It is intentional.

Also document that it's intentional that we don't have a default case, and add a comment to the source code to that extent.

Delete the existing exclusion, which allowed unwritten fields.  We don't actually have any unwritten fields right now.