Change constants from _COLUMN to _FIELD
Currently, "column" and "field" are the same thing.  That is, if you want to edit the "status" field, you know it's the column corresponding to the constant value.

Going forward, I want to make the columns more configurable: users should be able to rearrange columns, hide them, etc.  But current users of these constants are really referring to the fields.  Make the name accurate and free up the COLUMN names to be added later.