More protection against shutdown
Add several more "isDisposed()" checks.  Generally, it's not like we have to worry about other threads coming along and disposing the items.  Pretty much the only time it happens is when the app is being shut down.  But it definitely happens then, and if you tried to shut down the app while it's still fetching data, it could get hung up and not exit cleanly.

These changes probably don't guarantee a clean exit, but they make it pretty likely.