Modify isSameFile() somewhat
I might not have noticed previously that java.nio.file.Files.isSameFile() expects both of its arguments to exist on disk.  If you pass it a Path that doesn't exist, it throws an exception.

That's fine, in that we catch the exception, but we do report it.  I want a version that understands one of the files may not exist, and that that is fine, and expected, and no cause for an error report.

So there's now a slight distinction between the two arguments.  The first one, we expect to exist; if it doesn't, no calamity, but we will report an exception.  The second one, we do an explicit notExists() check first, and simply return false if it doesn't exist.

Update the comments for this, and remove the no-longer-accurate annotation.