Add currentLocationOf to keep episode db up to date
When we move files, the items may still stay on the table (depending on whether "delete row after move" is set or not).  If it stays in the table, we need to update it for its new location.  We do so with some code in ResultsTable, but this method (not used as of this commit) will help do it better.