Create updateTableItemAfterMove, using currentLocationOf
What we had been doing was, when the user clicked the move button, we'd move all the files, and then after all changes were done, we'd use refreshAll() to re-render every cell in the table (whether the file was moved or not).

Since the time we added that code, though, we have added a finishMove() callback, called after EACH move.  So, we can update the table item as soon as its move completes, and we don't need to refresh all at the end.

But, we can still do better.  We don't need to do a "full" refresh.  Moving a file does not change the proposed destination, because it doesn't change the episode information.  It also doesn't change (for now, at least) the status or the check box.

We also now do the work in a synchronized block, so the item is locked.  And, add a significant comment explaining what's going on.