Change replaceAll to replaceDestinations
The method "replaceAll" was properly named, and for how it was being used, did have to refresh everything.

As of the previous commit, when the "current" location changes due to the file being moved, we no longer call refreshAll().  We now have other code for that.

So, after that, the only remaining uses for refreshAll were to replace the proposed destination field.  So, change the method to only do that, and also change the name to keep it accurately named.

Also change it to now rely on currentLocationOf, and to be ready to handle the situation where the file has been moved out from under us.  And we no longer need to update the episode map.  Calling currentLocationOf() does it for us.