Tiny changes to ResultsTable
More checking for table being disposed, this time around creating the combo box.

Change setColumnDestText to take no arguments, and get the column itself.

Change updateUserPreferences to not take the UserPreferences, since its a singleton and we already have it as a static variable.

There were two signatures for setCellImage: one took an image, the other took a status.  Rename the one that takes a status, to "setCellStatus".

Add javadoc to three methods.