Make testExistingAncestorSymlinks a unit test
It was defined, but not annotated with @Test.  I think the reason was, since it uses symbolic links, it can't run on Windows.  Use the Environment class to check the platform, and do nothing on Windows, but otherwise run.

To see the actual (limited) extent of this change, view the commit with ?w=1.