Have checkstyle enforce Javadoc requirements
But our test files don't have Javadoc, and I don't want to add it; at least, not now.  We can suppress certain checkstyle rules by using a "SuppressionFilter", which we haven't had before.

So, add a "SuppressionFilter", a new file in etc/nochecks.xml.  It specifies to not apply the JavadocMethod rule to the test files, and also excludes two other files which are due for some major changes, and therefore not worth tweaking.