Do download outside of synchronized block.
In getShow, when we don't yet have a matched show, we know that the application will need to download the information.  But we don't want to start a download if one is already in progress.  So, we want to check if the showName already has any existing listeners; if so, that indicates that another file has already initiated the download.

When we want to check whether or not the object has listeners, we use a synchronized block.  It's likely that we'll be looking up multiple episodes of the same show, and they will be processed very closely to each other, and we could get race conditions.

However, there is no need to actually DO the downlod inside that synchronized block, which is what we were doing.  In fact, putting it in the block makes a deadlock.  Move the call to downloadShow outside of the synchronized block.