Eliminate inner class from ShowName
QueryString previously was a static inner class in ShowName.  Move its functionality up to ShowName.

I guess the whole distinction was kind of a dumb idea.  Here's the dumb idea:

If we get two files like:
  "The Office S05E07.mp4"
  "the.office.s06e20.mkv"

... I was thinking we should have separate ShowName objects for each, because the substrings we extract as the found show names are not identical:
  "The Office"
  "the.office"

However, their "query strings" ARE identical.  So that's what the previous setup did: had them have separate ShowName objects while sharing a QueryString object.

I realized that was pretty pointless.  I do still think it's important to retain the actual found show; if something goes wrong with "The Office S05E07.mp4", we don't want to say we had a problem parsing "the.office", because that string does not even appear in the filename!  So, we do maintain the original, untouched string.

But, we maintain it in FileEpisode.  For the work we're doing here, we don't need to retain it.

Especially since, as I was reminded in a recent commit, the whole point is to NOT process both ShowNames.  Since they have the same query string, we only need to process one of them; that's why we had the query string in the first place.  So, to say either "The Office" or "the.office" failed is wrong, anyway.  It's really the query string that failed.  If we do mention one or the other, it's only as an *example* of a filename that gave us this problem.  So, we do still keep an "example" found name along with the query string.  But we don't need separate show names, which means we'd have a 1-to-1 mapping between ShowNames and QueryStrings, which means they can just be the same object.