Add mapShowName, for clarity
We maintain a mapping between the "found show" (that is, the substring of the filename that we think refers to the series name), and the ShowName object that we use to represent it.  When we first encounter a new "found show", we create the mapping, and then later, we look it up and use it again.

These two usages should be distinguished.  They are really not the same thing.

Create a "new" method, mapShowName, which can be used to create the mappings, though if another file has already been processed which had an identical "found show", then the mapping will already exist, and we use that, instead.  But "new" is in quotes because the implementation of the "new" method is identical to the existing method.

Then, the existing method, "lookupShowName", will really only be used in cases where we want a "lookup", i.e., where we feel certain the mapping should already have been established, and we're just retrieving it.

In the unexpected case where we call lookupShowName and yet DON'T find a mapping... well, I guess we'll go ahead and create it, anyway.  What else could we do?  We could call it an unrecoverable error, but it likely would be recoverable.  And we'd prefer not to crash if we don't have to.

So, the "new" implementation of "lookupShowName" is the same as the old one, and the same as the new method.  The only difference is, we will log an error if we can't find the mapping.  And, of course, if I see that error come up in practice, I'll look into this further.

We call "mapShowName" from FilenameParser; that's where we extract the "found name", and we want to immediately create the mapping, if it doesn't already exist.

There are two other callers of "mapShowName", but they are for testing only.  One lives inside the test package, and the other is only called from within the test package.