Clean up when and where we encode characters
Having to encode characters is a special, irritating fact of life for things like getting data from the internet.  But it should be pushed out to correct scope.  We should not have been encoding the "query string".  The query string is what we want to query for, and that has nothing whatsoever to do with the fact that the actual query is done by a URL get.  We'll encode the characters to do the URL get, but when we're constructing the URL.

Likewise, if we need to change some characters to get the XML library to parse it correctly, do that immediately before giving it to the XML parser, and don't change the original text.

Not encoding the query string means we don't need to decode it.  In fact, the two decode methods are both unused at this point, but we'll keep them around, in case they're useful later.

Simplify some other details of getting the URL in HttpConnectionHandler.java, as well.  It should not be doing any encoding or decoding.  It should be able to assume that the URL it is given is already as it should be, and that it can return the text exactly as it got it.  The caller can worry about transforming the text.