Clean up HttpConnectionHandler some
Create private method getInputStream to make it clearer that all that work in the middle of downloadUrl was just creating the proper type of InputStream.  Rearrange it some so hopefully it's clearer still.

Now that we have getInputStream, it's simpler to use try-with-resources for the InputStream.  That means we can get rid of the finally block.