Use trimFoundShow
The process of finding a show name basically goes through three steps:
(1) the actual substring of the filename is called the "found" or "filename" show name
(2) the found show name is massaged into a query string, which we send to the provider
(3) the provider sends us back a list of actual show names, of which we choose one

So, the query string (2) is used to obtain the actual show name, and the actual name (3) is, of course, used for renaming the files.

So then what is the found name (1) used for?  It could just be a temporary local variable on the way to creating the query string.  But it is used for communicating with the user, and debugging.  If something goes wrong, we don't want to display the query string, because it's likely the filename does not actually, literally contain the query string.  We hold onto the "found name" to be able to refer back to where we got started.

Therefore, we should not generally "clean up" the found name.  The whole point is that it appears literally in the filename.

However, trimming it is different.  The substring that results from the trim STILL does appear literally in the filename, and in fact, it's more what anyone would expect to see being processed.

This commit introduces a huge number of diffs because FilenameParserTest had a large number of individual tests that verified the "filenameShow", and the expected value needs to be changed in each case.  This shouldn't be taken to indicate that this is a large change.

It's possible we could avoid NEEDING to trim in the first place, by tweaking the regular expressions, but this is far simpler and safer.

Going forward, we will be using the found show name a bit more, in giving the user the opportunity to "correct" our choices and generally have more control over how the new name is chosen.