Replace getFoundName with getExampleFilename
This commit doesn't change any behavior, it is just changing code to hopefully be more clear.

If we have two ShowNames:
 "The Office"
 "the.office"

... they will be separate objects, because their "foundName"s are not identical.  But their query strings ARE identical and because of that, one of them will be passed to TheTVDBProvider.java, and the other will not.  And there's no way to say which; it'll just be whichever ShowName happens to be created first.

So if there's a problem querying with that query string, we shouldn't exactly say we had a problem with one or the other.  We had the same problem with both.

So, our error message should not include "the found name", because that's misleading.  It's much better represented as an "example filename".

This commit also changes a couple of other methods in ways that do not affect the behavior of the program.