Make checked column wider
On OS X, it seems like the width of the column is the width IN ADDITION TO the size of the checkbox, so I made it very small.  But it doesn't work the same way on Linux.  Maybe we should make the default platform specific.  Unless or until I do that, it's better to be a little big on OSX, than too small on Linux.