Move Java source files into java dir
Moving towards a Maven/Gradle default setup, which requires the main Java files be not just in src/main, but src/main/java.

There's little sense in moving "towards" a different setup; either it conforms or it doesn't.  It just involves moving so many files that I'm spreading it out over a few steps.