Update run scripts
Recently, I created run-ubuntu.sh, basing it off of the existing scripts.  But in so doing, I found there was stuff that could be done better in the existing scripts.  Update them too, based on experience using them.

In order for the Java files, in development (i.e., running from class files rather than a jar) to find their resources, it seems the user.dir (working directory) needs to be a particular place.  So, allow the script to be run from any working directory, but for actually launching the JVM, make sure to cd into the right directory first.