Change launch4j properties file name
Since users are free to install launch4j wherever they want, we don't hard-code it into the build.xml file.  We rely on a properties file.  We also provide a default properties file, in etc.  The instructions on the github wiki say to copy that file to "build.properties" and potentially edit it.  And then the build.xml file looks for "build.properties".

But all it has in it is launch4j.  Use the name launch4j.properties, instead.  Forget about the "template" part, just call the checked in file "launch4j.properties".

Will need to edit the wiki as a separate step.