Don't delete dist build dirs
It's nice to clean up after ourselves, but I don't think the deletion was ever about that.  It was because we were re-using the same directories for different platforms.  Now that we no longer do that, don't delete the files.

It is helpful to be able to see what was created.