Use separate build variables for main and test
The object files for the main build should be going to a different location than the test files.  And they will be.  To try to keep each commit simpler, this just introduces two variables to build.xml -- build.main and build.test -- and tries to change things to use them in the right way.  But as of this commit, they have the same value, so they get blended together anyway.

There are a few other tiny changes, like "init" no longer makes the build dir; we need the build dir if and only if we're compiling, so we can just make it part of the compile task.  For the dist directory, there's no reason to build it until we're building a dist.