Rename swt jar files
We took the published SWT files, and to check them into jars/main, renamed them to simpler names that corresponded exactly with the distributable file (exe or zip) that we wanted to generate.

In the build.xml file, when packaging the files, the correct library was found by combining the "platform" attribute into a small template.  With the real files names, the "platform" doesn't match as well.

So, introduce a second attribute, "swtid", to identify the precise string that we should look for in the library name.

Also update the library names in etc/run-scripts