Add support for three-digit seasons (#357)
For certain more obscure formats, it's too tricky, but for the most common formats, which have clear delimiters, it's no problem to accept three-digit seasons.

We were given an example of a show that actually has over 100 seasons: House Hunters International.  Use it to add tests to confirm we are parsing such filenames correctly.

Also tested running the app with the example file.  It did work.  But it took a really, really long time to parse all the episodes.