Build outside of dist, include application icon, more
Don't build the bundle inside the dist directory; use another subdirectory of ${build}, and make it platform-specific, so that the 64-bit build does not overwrite the 32-bit one.

In order to include the application icon, it needs to be found by ant at a specific location with a specific name.  So, define that path, copy the icon to that location, and add that location to the .gitignore file.

Then, once the application is built in build.bundle, create the zip file directly into the dist directory.

Also change the "name" of the project; it had been "TVRenamer build script"; make it simply "TVRenamer" so that it names the app.  Then use ${ant.project.name} in place of TVRenamer in the bundling.