Update table when API discontinued
If we get a "not found" error from the server while trying to look up show options, we assume it's because the API we're using has been discontinued.  We put up a nice dialog box about it.  But we don't exit, and the files that we were unable to look up were stuck in the state where it said they were "Downloading...".

Change it instead to display the DOWNLOADING_FAILED message, "Downloading show listings failed".