Notice when provider times out
Previously, if we tried to download information from thetvdb.com, and it is down and doesn't respond, we gave a failure message that was no different than if the show was not found.  Try to notice when there is a SocketTimeoutException, and give a different replacement text that says that we timed out.