Simplify the logic around copy-and-delete
When copyAndDelete returns false, we can immediately return false; in the previous source code, there's nothing else we would do, we just delay the return for no good reason.

And we don't have to set actualDest = destPath in the success case, if we just initialize actualDest to destPath in the first place.