Reduce some verbosity in deleteFile
If we try to delete a file, there are likely numerous technical reasons why it could fail.  So if it does fail, our code was printing the exception, and the stack trace, to the console.  The simplest, most obvious way a file-delete could fail is that we do not have the required access to delete it.  When that is the case, Files.delete() throws a specific exception, of type AccessDeniedException.

If a file-delete fails due to an AccessDeniedException, we don't need to see the stack trace or the message.  We know what it means.  Give a shortened, meaningful warning, and go on.