Add a test that tries to move a file which can't be moved
Create an episode file, then set it to read-only, and more importantly, set its parent directory to read-only.  Then try to move it.  It should fail cleanly.  We should be able to verify that the original file is still in place, that the proposed destination file wasn't created, and that the FileMover.call() method returned false.

In order to help set the parent directory to read-only, find the parent directory, store it in an instance variable, and abort if it is null.