Reduce findbugs exclusions
We use findbugs to analyze the class files and report on potential bugs.  We want a clean bill of health every time.  Sometimes, it complains about things that we've done intentionally.  When that happens, we use exclude.xml to tell it not to consider those things bugs.

I guess there were a couple of things in ResultsTable.java that it was complaining about, that I didn't want to address at that time, but which since I've addressed.  So, we no longer need exclusions for those things.