Add timestamp check to move testing
When we move a file, the current behavior is to always update the timestamp after it is moved.  Although I personally am not sure I 100% agree with that behavior, I still want to know if it has stopped working, so check the timestamp.

There is not a extremely precise expected value.  The expected value is whatever the System time was at the instant we decided to write the timestamp.  But we can get the System time value as soon as the move is completed, and expect it to be pretty close.  Set a threshold of 1000 milliseconds for the test to pass.  (If it turns out anyone wants to run these tests on machines that are slower and busier, and this is causing false negatives, it'd be fine to increase the threshold.).