Add FileEpisode test for three-digit season
A previous commit to FilenameParser implemented the fix for #357, "Can't handle Seasons with 3 digits", and it included testing the parsing.  This commit adds a test that the output works, and in particular, using the "%0s" control string.  That zero-pads the season to two digits.  In the case of a three-digit season, it should just include all three digits; padding does not imply truncating.

Everything was working as it should; this is just a test to confirm it and make sure it doesn't get broken.