Move the call to finishProgress
Notify the observer of finishProgress in tryToMoveRealPaths, the calling function, rather than in doActualMove or copyAndDelete.  This way we only have the code one time, instead of three.

Note, we consider it a "failure" in cases where (a) we copy the file successfully, but can't delete the original, and (b) copy/move the file successfully, but cannot set the timestamp of the new file.

We may need a richer set of statuses than simply success/failure.