Have UserPreferences keep track of the dest dir Path
The user specifies the destination directory (labeled the "TV Directory" in the dialog box: the directory to which we will move the files) with a String, and we need that String to display the "Proposed File Path" column, and it is serialized (to prefs.xml) and deserialized as a String.

But the point of it is that it represents a directory on a disk.  So it needs to be converted to a Path object to be used.

Previously, classes that used the UserPreferences would obtain the value as a String, and convert it themselves.  Change it so that UserPreferences maintains a Path object.  (Note, whenever we take a value for "TV Directory", whether at startup or because the user has edited it, we verify the directory, which requires creating a Path object.  It's just that previously, we'd drop it on the floor when we were done with it.  Now we're holding onto it.)

Add method, getDestinationDirectory().  This is different from getDestinationDirectoryName() in two ways.  One, of course, it does not return a String, but a Path object.  The other is, it only returns a valid Path object when "move" is enabled; if "move" is disabled, there is no "destination directory", and so it returns null.

We will continue to use the String for serialization and deserialization.  Therefore, the Path instance variable is marked transient.

Other classes that use UserPreferences, like the ResultsTable, actually only want the String, and so that API method is maintained.

With this change, we no longer need the versions of checkForCreatableDirectory and existingAncestor that take String arguments (in FileUtilities).  Remove them, and remove the code that tested them.  Then, that class has nothing remaining that takes Strings and creates Paths from them; it always gets the actual Path.  So we can eliminate the import of the Paths class, which basically just exists to create Path objects from Strings.