Don't depend on "resolve" for compile.test
In order to compile anything, the libraries need to be downloaded, which is done via "ant resolve".  But that really is basically a one-time thing.  Once you download them, they're there.

Running "ant resolve" again does not take a whole lot of time: it checks that all the right versions are there, and then, even if anything is missing, it will go to the cache before actually downloading from the itnernet.

Nevertheless, it's hardly instantaneous, and it throws a lot of output on the screen.  So, I did not make "compile" depend on it.  Instead, I added a check that the lib directory is there; if it's not, then we know we need to run "ant resolve", so advise the user as such.  If it's there, it's probably ok.

I did initially make "compile.test" depend on it, but in retrospect, I don't like that, either.  Change "compile.test" to be just like "compile": check for the lib directory, but that's it.

Move the "resolve" dependency down to "test".  So, that still remains a good way for people to get everything all set up.  If you clone the repository and run "ant test", it will download libraries and compile everything and run tests, and if it passes, you know you're good to go.