Do work inside getMoveToPath
Previously, FileEpisode had two methods: a private one called getMoveToDirectory, which composed a String from the user-specified destination directory, replacement template, and the information we've found about the file; and a public one caleld getMoveToPath, which basically called getMoveToDirectory and turned the result into a Path.

This made some sense since we work with Strings a lot: all the various components are Strings, and we need a String to put into the ResultsTable.

Still, the point ultimately is to move the file, so the actual Path is pretty important; plus, using Path.resolve() seems better than concatenating Strings with FILE_SEPARATOR_STRING.

We still need the Strings.  But switch it around: create the Path object, and use toString() to get the String version.