Add "jtest" target
As the comment says, "jtest" could stand for Java test -- or it could stand for "just test".  Compile any uncompiled files and run the tests, but without doing a "clean" or a "resolve" first.

The "test" target is changed to depend on "jtest", so it just does a clean, resolve, and then the same thing as jtest, which is the same behavior that "test" had before this commit.

The "jtest" target is intentionally not given a "description", because it should not be considered a "primary" task.  It's more like an undocumented feature.