Improve fail-to-move message from the UI
The UI is responsible for kicking off the "actual move" of the files, and needs to update its table after the move has finished.  When a move fails, it logs a message to the console about the failure.

Previously, it just printed the TableItem, which has a very uninformative implementation of toString.  Instead of printing the item, print the "current file field" of the item.

Note that, in practice, this is completely redundant.  It follows a very similar message that is printed by the FileMover.  But I think it's good for each class to log its own exceptional situations.