Add read-only test for MoveRunner
Previously added a read-only test that used FileMover directly, which is a lot simpler because there are no threads/callbacks.

Now test it with MoveRunner, which is much more similar to what the real program does.

It's not much different from testFileMoverCannotMove(); we expect the same things.  We have to handle exceptions here, that would not come up in testFileMoverCannotMove().  Really the main thing to check here is that we are still getting failure results sent to the ProgressObserver.