Cleanup of MoveTest
In a previous commit, I introduced a (top-level) abstract class that implemented the ProgressUpdater interface, which is used by MoveRunner; what I actually wanted was one that implemented ProgressObserver, which is used by FileMover.  That's the one that really tells us when the file move is complete.

Instead of a top-level one, though, replace NoOpProgressUpdater with an inner class in MoveTest that implements ProgressObserver.

Also, move the "assertMoved" out of the finish method entirely, and put it into the main thread.

Check the return status of finishProgress, which we get via the CompletableFuture.

Finally, FileMover may write out some info messages, which I don't think we want to see during the testing.  So I had already put in a line to dial down the verbosity before we start the tests.  But I commented it out to confirm the difference, and accidentally committed it that way.

Un-comment it.