Bug in FailedShow
I clearly made a mistake in the previous commit.  There is code in the FailedShow constructor that is intended to find if there was SocketTimeoutException anywhere in the chain of Throwables that led to the given TVRenamerIOException.  But a crucial line was left out: the one that changes the variable to be its parent.  Kind of seems like it would be an infinite loop in all cases, the way it is.

Actually, it still could be an infinite loop, if it's possible for a Throwable to be its own parent, or to have any kind of a cycle.  But I'm going to assume that's not the case.