Add TestUtils file, with createFile method
Add method to ensure the existence of a file, in a given directory, with a given name.  If the file exists already, the method simply returns true.  If it doesn't exist, the method tries to create the directory, create the file, and write some relatively arbitrary content into the file -- the filename itself, encoded in ISO-8859.  But the content is not the point.