Keep MoveStatus in FileMover
Add three new values to the MoveStatus enum: ALREADY_IN_PLACE, MISNAMED, and COPIED.  Now it is sufficient to represent all of the states that a FileMover may go through as it tries to move a file.

Change FileEpisode.setRenamed to take a status as an argument, rather than assuming a status of RENAMED.  Add javadoc about the various statuses that we may expect.  Note that FileEpisode.fileStatus remains a "write-only" instance variable.  We should use it in the near future, but as of now, we don't.

In FileMover, add a status variable, and update it in all the places where the status of the move changes.