Move block where we check destination
When we move a file, we check to make sure the destination where the file has been moved to, is where we actually intended and expected.

This check may well be superfluous.  It is probably impossible for the move to go to the wrong place.  But the Java library call to move a file returns a destination, so it seems to make sense to check it.

However, that's not the case in a copy-and-delete.  In that case, it's absolutely impossible to get a different destination, because we don't even get a return value.

So, move the block of code which checks that the actual destination is equal to the expected one, up into the block where we do the normal "move".  Eliminate pointless local variable.

Also add a log message for the copy-and-delete case where we successfully copy the file but then can't delete the original.