Inform user of move failures (#376)
Prior to this, when we failed to move one or more files, there was no obvious indication given to the user.

(When a file IS successfully moved, the value displayed in CURRENT_FILE_FIELD changes, and also, if "Delete Rows After Move" is checked, the row is deleted. So, the fact that neither of those things happens with a failure is SOME indication of the failure, but it's not obvious.)

I'm using a modal dialog box, which is kind of aggressive.  I think it's justified in this case, but I want to make sure we don't show more than one per move operation.  As results come in, asynchronously, we store the failures into a data structure, and then once all moves have completed, if there were any failures, we construct the text for the dialog box and present it to the user.

This provides the basic functionality.  There is clearly more that could be done.